+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## [0.0.1] - TODO: Add release date.
-
-* TODO: Describe initial release.
+## [1.1.0] - 25.04.2018
+* Added an option which allows to wrap the image in an AspectRatio
+* Added an option to lock scaling
+## [1.0.0] - 20.04.2018
+* Initial release